+++ src/deshyperneat/node.rs
@@ -62,7 +62,7 @@ impl NeatNode for Node {
     fn distance(&self, other: &Self) -> f64 {
         let mut distance = self.core.distance(&other.core);
         distance += 0.8 * self.cppn.distance(&other.cppn);
-        distance += 0.2 * ((self.depth - other.depth) as f64).abs().tanh();
+        distance += 0.2 * (self.depth as f64 - other.depth as f64).abs().tanh();
         distance
     }
 }
